reorganize styles scripts to use src directory
