+++ .env.development
@@ -14,6 +14,7 @@ SERVICE_EMAIL_PDF_SUBHEADING='Sub Heading'
 SERVICE_EMAIL_FROM="form-builder-developers@digital.justice.gov.uk"
 SERVICE_EMAIL_SUBJECT="Subject"
 SERVICE_EMAIL_BODY="Email body"
+AUTOCOMPLETE_FIXTURE=countries
 
 # Details from the fb-acceptance-tests-repo
 SUBMISSION_ENCRYPTION_KEY='52905141-4738-4cce-8bd6-633c970c'
