+++ app/javascript/packs/govuk.js
@@ -1,3 +1,6 @@
 require("govuk-frontend/govuk/all").initAll()
 require.context('govuk-frontend/govuk/assets/images', true)
-import "./stylesheets/govuk.scss"
+const accessibleAutocomplete = require("accessible-autocomplete")
+window.accessibleAutocomplete = accessibleAutocomplete
+
+import "../styles/govuk.scss"
