+++ app/javascript/src/__application.js
