+++ app/javascript/src/analytics.js
